moved ErrorComponents out to components folder
